requirements.txt
Flask
gunicorn
openai
groq
pandas
matplotlib
python-dotenv
tabulate
openpyxl
